requirements.txt
Flask
gunicorn
pyjwt
requests
cryptography
authlib
six
datetime
python-dotenv
sqlalchemy
flask-login
flask-migrate
psycopg2-binary